+++ src/util/foray_pipelinebuilder.cpp
@@ -72,7 +72,7 @@ namespace foray::util {
         if(mColorBlendAttachmentStates == nullptr)
         {
             blendAttachmentStates.resize(mColorAttachmentBlendCount);
-            for(int i = 0; i < mColorAttachmentBlendCount; i++)
+            for(int32_t i = 0; i < (int32_t)mColorAttachmentBlendCount; i++)
             {
                 blendAttachmentStates[i].blendEnable    = false;
                 blendAttachmentStates[i].colorWriteMask = VK_COLOR_COMPONENT_R_BIT | VK_COLOR_COMPONENT_G_BIT | VK_COLOR_COMPONENT_B_BIT | VK_COLOR_COMPONENT_A_BIT;
